Backend/requirements.txt
aerich==0.9.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncclick==8.3.0.7
asyncio==4.0.0
asyncpg==0.30.0
attrs==25.4.0
bcrypt==5.0.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
dictdiffer==0.9.0
disutils==1.4.32.post2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.1
fonttools==4.60.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpx==0.27.2
idna==3.11
iso8601==2.1.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
motor==3.7.1
multidict==6.7.0
numpy==2.3.4
odmantic==1.0.2
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.9.0
pymongo==4.15.3
pyparsing==3.2.5
pypika-tortoise==0.6.2
python-dateutil==2.9.0.post0
pytz==2025.2
ratelimit==2.2.1
redis==7.0.0
regex==2025.10.23
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
threadpoolctl==3.6.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uuid==1.30
uvicorn==0.38.0
yarl==1.22.0
